moving fetch of visual element to the same time as concept fetch
